Move situm to provider
