+++ index.html
@@ -132,6 +132,6 @@
       </div>
       <div class="loading__text">Loading&hellip;</div>
     </div>
-    <script type="module" src="/src/index.jsx"></script>
+    <script type="module" src="/src/index.tsx"></script>
   </body>
 </html>
