--- src/index.jsx
@@ -1,21 +1,0 @@
-import { IconContext } from "react-icons";
-import { createRoot } from 'react-dom/client';
-import { Provider } from 'react-redux';
-
-import './favicon.ico'
-import './styles/index.scss'
-import './i18n';
-import App from './components/App';
-import { store } from './store';
-
-const root = createRoot(document.querySelector("#app"));
-root.render(
-  <Provider store={store}>
-    <IconContext.Provider value={{className: 'react-icons'}}>
-      <App/>
-    </IconContext.Provider>
-  </Provider>
-);
-
-// Hide the loader.
-document.querySelector(".loading").style.display = "none";
