+++ src/store/index.ts
@@ -1,25 +1,15 @@
 import { configureStore } from '@reduxjs/toolkit';
 import type { RootState } from './types';
 
-// Import all slices
-import styleReducer from './slices/styleSlice';
-import uiReducer from './slices/uiSlice';
-import layersReducer from './slices/layersSlice';
-import sourcesReducer from './slices/sourcesSlice';
-import errorsReducer from './slices/errorsSlice';
-import mapViewReducer from './slices/mapViewSlice';
-import debugReducer from './slices/debugSlice';
+// Import merged slices
+import styleCoreReducer from './slices/styleCoreSlice';
+import uiCoreReducer from './slices/uiCoreSlice';
 
 // Configure the store
 export const store = configureStore({
   reducer: {
-    style: styleReducer,
-    ui: uiReducer,
-    layers: layersReducer,
-    sources: sourcesReducer,
-    errors: errorsReducer,
-    mapView: mapViewReducer,
-    debug: debugReducer,
+    styleCore: styleCoreReducer,
+    uiCore: uiCoreReducer,
   },
   // Enable Redux DevTools in development
   devTools: process.env.NODE_ENV !== 'production',
